pyproject.toml
[project]
name = "airflow-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow-providers-google>=17.1.0",
    "apache-airflow-providers-http>=5.3.3",
]

[tool.ruff]
line-length = 120

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
